Fixed logging. Exception with empty email_list

+++ src/mailing.py
@@ -24,29 +24,36 @@ def json_to_excel(json_data):
 def execute_mailing():
     driver = DBDriver()
     email_list = driver.get_email_list_for_sending()
-    json_data = driver.get_receipts()
-    excel_filepath = json_to_excel(json_data)
-    msg = MIMEMultipart()
-    msg['Subject'] = "Mailing list from the taxBot according to your request (EXCEL file)"
-    msg['From'] = EMAIL_LOGIN
-    msg['To'] = ', '.join(email_list)
-    body = "This is an automated email"
-    msg.attach(MIMEText(body, 'plain'))
-    part = MIMEBase('application', "octet-stream")
-    with open(excel_filepath, "rb") as file:
-        part.set_payload(file.read())
-    encoders.encode_base64(part)
-    part.add_header('Content-Disposition', f'attachment; filename="{os.path.basename(excel_filepath)}"')
-    msg.attach(part)
-    os.remove(excel_filepath)
-    try:
-        smtpObj = smtplib.SMTP('smtp.gmail.com', 587)
-        smtpObj.starttls()
-        smtpObj.login(EMAIL_LOGIN, EMAIL_PASSWORD)
-        smtpObj.sendmail(EMAIL_LOGIN, msg['To'], msg.as_string())
-        smtpObj.quit()
-        _logger.error("E-mail has been sent successfully. STATUS_FAIL")
-        return STATUS_OK
-    except smtplib.SMTPException:
-        _logger.error("E-mail has not been sent. STATUS_FAIL")
-        return STATUS_FAIL
+    if email_list is None:
+        _logger.error("E-mail list is empty. Check records in database")
+        raise TypeError(
+            "Variable 'email_list' has type None, but must be of type 'list'. "
+            "This error possibly related with database"
+        )
+    else:
+        json_data = driver.get_receipts()
+        excel_filepath = json_to_excel(json_data)
+        msg = MIMEMultipart()
+        msg['Subject'] = "Mailing list from the taxBot according to your request (EXCEL file)"
+        msg['From'] = EMAIL_LOGIN
+        msg['To'] = ', '.join(email_list)
+        body = "This is an automated email"
+        msg.attach(MIMEText(body, 'plain'))
+        part = MIMEBase('application', "octet-stream")
+        with open(excel_filepath, "rb") as file:
+            part.set_payload(file.read())
+        encoders.encode_base64(part)
+        part.add_header('Content-Disposition', f'attachment; filename="{os.path.basename(excel_filepath)}"')
+        msg.attach(part)
+        os.remove(excel_filepath)
+        try:
+            smtpObj = smtplib.SMTP('smtp.gmail.com', 587)
+            smtpObj.starttls()
+            smtpObj.login(EMAIL_LOGIN, EMAIL_PASSWORD)
+            smtpObj.sendmail(EMAIL_LOGIN, msg['To'], msg.as_string())
+            smtpObj.quit()
+            _logger.info("E-mail has been sent successfully. STATUS_OK")
+            return STATUS_OK
+        except smtplib.SMTPException:
+            _logger.error("E-mail has not been sent. STATUS_FAIL")
+            return STATUS_FAIL
